requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=py37haa95532_0
addict=2.4.0=pypi_0
albumentations=0.5.2=pypi_0
anyio=2.2.0=py37haa95532_2
argon2-cffi=20.1.0=py37h2bbff1b_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py37haa95532_0
blosc=1.21.0=h19a0ad4_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2022.2.1=haa95532_0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2021.10.8=py37haa95532_2
cffi=1.14.5=py37hcd4344a_0
chardet=4.0.0=py37haa95532_1003
charls=2.1.0=h33f27b4_2
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.7=py37h71e12ea_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37he774522_0
dask-core=2.17.0=py_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecated=1.2.13=pypi_0
entrypoints=0.3=py37_0
et_xmlfile=1.1.0=py37haa95532_0
freetype=2.10.4=hd328e21_0
giflib=5.2.1=h62dcd97_0
google-auth=1.21.3=py_0
google-auth-oauthlib=0.4.3=pyhd3eb1b0_0
graphviz=2.38.0=4
grpcio=1.35.0=py37hc60d5dd_0
h5py=2.8.0=py37hf7173ca_2
hausdorff=0.2.6=pypi_0
hdf5=1.8.20=hac2f561_1
hiddenlayer=0.3=pypi_0
humanize=3.13.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imagecodecs=2021.1.11=py37h5da4933_1
imageio=2.9.0=py_0
imgaug=0.4.0=pypi_0
importlib-metadata=3.7.3=py37haa95532_1
importlib_metadata=3.7.3=hd3eb1b0_1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=3.0.0=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter-packaging=0.7.12=pyhd3eb1b0_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37haa95532_0
jupyter_server=1.4.1=py37haa95532_0
jupyterlab=3.0.14=pyhd3eb1b0_1
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.4.0=pyhd3eb1b0_0
kiwisolver=1.3.1=py37hd77b12b_0
lcms2=2.11=hc51a39a_0
lerc=2.2.1=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.7=h2bbff1b_5
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libzopfli=1.0.3=ha925a31_0
lpips=0.1.3=pypi_0
lz4-c=1.9.3=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py37haa95532_0
markupsafe=2.0.1=py37h2bbff1b_0
matplotlib=3.3.4=py37haa95532_0
matplotlib-base=3.3.4=py37h49ac443_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.3.0=py37h46781fe_0
mkl_random=1.1.1=py37h47e9c7a_0
mmcv=1.4.6=pypi_0
msys2-conda-epoch=20160418=1
natsort=7.1.1=pyhd3eb1b0_0
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
nibabel=3.2.1=pypi_0
ninja=1.7.2=0
notebook=6.4.0=py37haa95532_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
opencv-python=4.5.1.48=pypi_0
opencv-python-headless=4.5.1.48=pypi_0
openjpeg=2.3.0=h5ec785f_1
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1m=h2bbff1b_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py37hd77b12b_0
pandoc=1.19.2.1=1
pandocfilters=1.4.3=py37haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pdf2image=1.16.0=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py37h4fa10fc_0
pip=22.1.2=pypi_0
piq=0.5.5=pypi_0
piqa=1.1.3=pypi_0
prefetch-generator=1.0.1=pypi_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.14.0=py37hd77b12b_1
py-opencv=3.4.2=py37hc319ecb_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.4=pypi_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyjwt=1.7.1=py37_0
pymupdf=1.19.6=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.10=h6244533_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.16=pyhd3eb1b0_1
pytorch=1.8.1=py3.7_cuda10.2_cudnn7_0
pytorch-metric-learning=0.9.99=pyh39e3cac_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37he774522_2
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyzmq=20.0.0=py37hd77b12b_1
qt=5.9.7=vc14h73c81de_0
regex=2022.3.2=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.17.2=py37h1e1f486_0
scikit-learn=0.24.1=py37hf11a4ad_0
scipy=1.6.2=py37h14eb087_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37haa95532_0
shapely=1.7.1=pypi_0
simpleitk=2.1.1=pypi_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py37haa95532_0
snappy=1.1.8=h33f27b4_0
sniffio=1.2.0=py37haa95532_1
sqlite=3.35.2=h2bbff1b_0
surface-distance-based-measures=0.1=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
terminado=0.9.4=py37haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.17=pyhd3eb1b0_1
tk=8.6.10=he774522_0
toolz=0.11.1=pyhd3eb1b0_0
torch-fidelity=0.3.0=pypi_0
torchaudio=0.8.1=py37
torchfile=0.1.0=pypi_0
torchio=0.18.71=pypi_0
torchmetrics=0.9.2=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.9.1=py37_cu102
tornado=6.1=py37h2bbff1b_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.13.3=pypi_0
xlrd=2.0.1=pyhd3eb1b0_0
xlwt=1.3.0=py37_0
xpinyin=0.7.6=pypi_0
xz=5.2.5=h62dcd97_0
yapf=0.32.0=pypi_0
zeromq=4.3.3=ha925a31_3
zfp=0.5.5=hd77b12b_4
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
